\name{variant.float.attribute}
\alias{variant.float.attribute}
\title{variant.float.attribute}
\usage{variant.float.attribute(vc,att)}
\description{No description available}
\value{@return the the INFO attribute for the given key.}
\arguments{
	\item{vc}{the variant}
	\item{att}{the INFO/Attribute}
}
