\name{variant.genotype}
\alias{variant.genotype}
\title{variant.genotype}
\usage{variant.genotype(vc,nameOrIdx)}
\description{return the genotype for a variant
}
\value{@return the given genotype}
\arguments{
	\item{vc}{the variant}
	\item{nameOrIdx}{the sample name (slower) or the 1-based sample index }
}
