\name{variant.string.attribute}
\alias{variant.string.attribute}
\title{variant.string.attribute}
\usage{variant.string.attribute(vc,att, split = TRUE)}
\value{@return the the INFO attribute for the given key.}
\arguments{
	\item{vc}{the variant}
	\item{att}{the INFO/Attribute}
	\item{splitString}{split strings using commas}
}
